Implemented Start Screen
Also added a small script where you can move from very first page to the start screen. I also noticed the project  direction was different from the one that was supposed to be. So now, I am working in right direction. That's the reason of all the documents that are attaching now in this commit.