Added game progress tracker and the functionality.
The functionality mean when game progress tracker goes to the end, there will be no more zombies to spawn. It is like a timer to let user know how far he is from the end of the level. Also worked on Zombie and ZombieSpawner scripts. Added an array for each spawn lines. This will make it easier to decided which zombies to be spawn in each line.